Create Profile tests and tweak method
Still doesn't work correctly
